# openapi.unified.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "TicketRemaster Unified API",
    "version": "2026-03-29",
    "description": "Unified OpenAPI document for the current TicketRemaster gateway-facing orchestrator API plus the external OutSystems credit service."
  },
  "servers": [
    {
      "url": "https://ticketremasterapi.hong-yi.me",
      "description": "Production gateway"
    },
    {
      "url": "http://localhost:8000",
      "description": "Local gateway"
    }
  ],
...
  "paths": {
...
    "/events": {
      "get": {
        "tags": ["Events"],
        "summary": "List events",
        "parameters": [
          {
            "name": "type",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "page",
            "in": "query",
            "schema": {
              "type": "integer"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Enriched events",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DataEnvelope"
                }
              }
            }
          }
        }
      }
    },
    "/events/{eventId}": {
      "get": {
        "tags": ["Events"],
        "summary": "Get event",
        "parameters": [
          {
            "name": "eventId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Event detail",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DataEnvelope"
                }
              }
            }
          },
          "404": {
            "$ref": "#/components/responses/Error404"
          }
        }
      }
    },
...
    "/marketplace": {
      "get": {
        "tags": ["Marketplace"],
        "summary": "Browse marketplace listings",
        "security": [
          {
            "GatewayApiKey": []
          }
        ],
        "parameters": [
          {
            "name": "eventId",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "page",
            "in": "query",
            "schema": {
              "type": "integer"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Listings",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DataEnvelope"
                }
              }
            }
          }
        }
      }
    },
    "/marketplace/list": {
      "post": {
        "tags": ["Marketplace"],
        "summary": "Create marketplace listing",
        "security": [
          {
            "BearerAuth": [],
            "GatewayApiKey": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MarketplaceListRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Listing created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DataEnvelope"
                }
              }
            }
          }
        }
      }
    },
    "/marketplace/{listingId}": {
      "delete": {
        "tags": ["Marketplace"],
        "summary": "Cancel marketplace listing",
        "security": [
          {
            "BearerAuth": [],
            "GatewayApiKey": []
          }
        ],
        "parameters": [
          {
            "name": "listingId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Cancelled",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DataEnvelope"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      },
      "GatewayApiKey": {
        "type": "apiKey",
        "in": "header",
        "name": "apikey"
      },
      "OutSystemsApiKey": {
        "type": "apiKey",
        "in": "header",
        "name": "X-API-KEY"
      }
    },
    "responses": {
...
      "Error404": {
        "description": "Not found",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorEnvelope"
            }
          }
        }
      },
...
    },
    "schemas": {
      "DataEnvelope": {
        "type": "object",
        "properties": {
          "data": {
            "type": "object",
            "additionalProperties": true
          }
        }
      },
      "ErrorEnvelope": {
        "type": "object",
        "properties": {
          "error": {
            "type": "object",
            "properties": {
              "code": {
                "type": "string"
              },
              "message": {
                "type": "string"
              },
              "status": {
                "type": "integer"
              },
              "traceId": {
                "type": "string"
              }
            },
            "required": ["code", "message"]
          }
        },
        "required": ["error"]
      },
...
      "MarketplaceListRequest": {
        "type": "object",
        "required": ["ticketId"],
        "properties": {
          "ticketId": {
            "type": "string"
          },
          "price": {
            "type": "number"
          }
        }
      },
...
    }
  }
}